Ember Route Authorization

+++ app/controllers/author/detail.js
@@ -1,11 +1,16 @@
 import Controller from '@ember/controller';
+import { inject as service } from '@ember/service';
 
 export default Controller.extend({
+  can: service('can'),
+
   actions: {
     deleteAuthor(author) {
-      author.destroyRecord().then(() => {
-        this.transitionToRoute('author.index');
-      });
+      if (this.get('can').can('delete author', author)) {
+        author.destroyRecord().then(() => {
+          this.transitionToRoute('author.index');
+        });
+      }
     }
   }
 });
