+++ app/controllers/book/detail.js
@@ -1,12 +1,17 @@
 import Controller from '@ember/controller';
+import { inject as service } from '@ember/service';
 
 export default Controller.extend({
+  can: service('can'),
+
   actions: {
     destroy() {
-      this.model.destroyRecord()
-        .then(() => {
-          this.transitionToRoute('book');
-        });
+      if (this.get('can').can('delete book', this.model)) {
+        this.model.destroyRecord()
+          .then(() => {
+            this.transitionToRoute('book');
+          });
+      }
     }
   }
 });
